requirements-reply-engine.txt
openai>=1.58.0
tweepy>=4.14.0
python-dotenv>=1.0.1
